Split Task.fallback(upon:to:) out to a separate file
